Move logic portion to HomeContainer for better separation of concerns
