+++ src/screens/HomeScreen.tsx
@@ -1,54 +1,17 @@
 /* eslint-disable import/prefer-default-export */
-import React, { useContext } from 'react';
+import React from 'react';
 import { View } from 'react-native';
 
 import HomeHeader from '../components/Home/HomeHeader';
 import { HomeRecentActivity } from '../components/Home/HomeRecentActivity';
-import { QuestionContext } from '../contexts/QuestionsContext';
-import { ResponseContext } from '../contexts/ResponseContext';
+import { BadgeTitleRewardText } from '../components/HomeContainer';
 import styles from '../styles/HomeScreenStyles';
-import { Home } from '../models';
 
-export interface BadgeTitleRewardText {
-  badgeTitle: string;
-  rewardText: string;
+interface HomeScreenProps {
+  badgeCompletedTextList: (BadgeTitleRewardText | undefined)[];
 }
 
-export const HomeScreen = () => {
-  const { questionState } = useContext(QuestionContext);
-  const { responseState } = useContext(ResponseContext);
-
-  const positiveResponse = responseState.items.filter(
-    response => response.answer === 'Yes',
-  );
-
-  const sortedResponses = positiveResponse.sort((a, b) => {
-    const date1 = new Date(a.createdAt);
-    const date2 = new Date(b.createdAt);
-    if (date1 > date2) {
-      return -1;
-    }
-    return 1;
-  });
-
-  const badgeCompletedTextList: (BadgeTitleRewardText | undefined)[] =
-    sortedResponses.map(response => {
-      const questionObject = questionState.items.find(
-        question => question.id === response.questionID,
-      );
-
-      if (questionObject === null || questionObject === undefined) {
-        return undefined;
-      }
-
-      const badgeCompletedTextObj: BadgeTitleRewardText = {
-        badgeTitle: questionObject.title,
-        rewardText: questionObject.rewardText,
-      };
-
-      return badgeCompletedTextObj;
-    });
-
+const HomeScreen: React.FC<HomeScreenProps> = ({ badgeCompletedTextList }) => {
   return (
     <View style={styles.header}>
       <HomeHeader />
@@ -56,3 +19,5 @@ export const HomeScreen = () => {
     </View>
   );
 };
+
+export default HomeScreen;
